requirements.txt
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
deepspeed==0.6.0
docker-pycreds==0.4.0
einops==0.3.0
filelock==3.9.0
ftfy==6.1.1
gitdb==4.0.10
GitPython==3.1.30
hjson==3.1.0
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
joblib==1.2.0
ninja==1.11.1
numpy==1.21.6
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==23.0
pathtools==0.1.2
Pillow==9.4.0
promise==2.3
protobuf==3.19.6
psutil==5.9.4
py-cpuinfo==9.0.0
python-dateutil==2.8.2
PyYAML==6.0
regex==2022.10.31
requests==2.28.2
sacremoses==0.0.53
sentry-sdk==1.14.0
setproctitle==1.3.2
shortuuid==1.0.11
six==1.16.0
smmap==5.0.0
termcolor==2.2.0
timm==0.5.4
tokenizers==0.12.1
torchmetrics==0.10.3
torchtyping==0.1.4
tqdm==4.64.1
transformers==4.18.0
typeguard==2.13.3
typing_extensions==4.4.0
urllib3==1.26.14
wandb==0.12.11
wcwidth==0.2.6
yaspin==2.3.0
zipp==3.11.0
